+++ src/components/survey/SurveySingleItemView/ResponseComponent/InputTypes/SingleChoiceGroup.tsx
@@ -14,6 +14,7 @@ interface SingleChoiceGroupProps {
   prefill?: ResponseItem;
   responseChanged: (response: ResponseItem | undefined) => void;
   languageCode: string;
+  showOptionKey?: boolean;
 }
 
 
@@ -174,6 +175,9 @@ const SingleChoiceGroup: React.FC<SingleChoiceGroupProps> = (props) => {
       })}
       key={option.key} >
       <div>
+        {props.showOptionKey ?
+          <span className="text-primary me-2">{option.key}</span>
+          : null}
         <input
           className="form-check-input cursor-pointer"
           type="radio"