+++ src/components/survey/index.ts
@@ -1,5 +1,9 @@
+import SurveySingleItemView from './SurveySingleItemView/SurveySingleItemView';
+import SurveyPageView from './SurveyView/SurveyPageView/SurveyPageView';
 import SurveyView from './SurveyView/SurveyView';
 
 export {
-  SurveyView
+  SurveyView,
+  SurveySingleItemView,
+  SurveyPageView,
 }